"validaciones"

+++ licoreria-toby/src/producto/producto.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Product } from './entities/producto.entity';
 import { Repository } from 'typeorm';
@@ -12,24 +12,42 @@ export class ProductoService {
 
 
 
-    crearProduct(producto: CreateProductoDto){
+    async crearProduct(producto: CreateProductoDto){
+
+        const existnombre = await this.productRepository.find(
+            {where: {nombre: producto.nombre}}
+        );
+        if(existnombre){
+            return new HttpException('El producto ya existe', HttpStatus.CONFLICT);
+        }
+        
         const crearProducto = this.productRepository.create(producto)
         return this.productRepository.save(crearProducto);
     }
-    getProduct(){
-
-        
-        return this.productRepository.find();
+    async getProduct(){
+        return await this.productRepository.find();
     }
+
+
     async getProductId(id: number){
-        return await this.productRepository.findOne({
+
+        const product = await this.productRepository.findOne({
             where: {
                 idproducto: id
             }
         });
+        if(!product){
+            return new HttpException('El producto no existe', HttpStatus.NOT_FOUND);
+        }
+        return product;
     }
 
-    deleteProductId(id: number){
+    async deleteProductId(id: number){
+        const foundproduct = await this.productRepository.findOne({where: {idproducto: id}});
+
+        if(!foundproduct){
+            return new HttpException('El producto no se encuentra', HttpStatus.CONFLICT)
+        }
         return this.productRepository.delete(id);
     }
     async updateUser(id: number, product: UpdateProductoDto){
